requirements.txt
aiohttp==3.9.3
aiohttp-retry==2.8.3
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.2.0
appnope==0.1.3
asttokens==2.4.1
attrs==23.2.0
blinker==1.7.0
catalogue==2.0.10
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
comm==0.2.1
dataclasses-json==0.6.3
dateparser==1.2.0
debugpy==1.8.0
decorator==5.1.1
deprecation==2.1.0
distro==1.9.0
emoji==2.10.0
executing==2.0.1
Faker==22.6.0
filelock==3.13.1
Flask==3.0.1
frozenlist==1.4.1
fsspec==2023.12.2
gotrue==2.1.0
gunicorn==21.2.0
h11==0.14.0
httpcore==1.0.2
httpx==0.25.2
idna==3.6
ipykernel==6.29.0
ipython==8.20.0
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.3
joblib==1.3.2
jsonpatch==1.33
jsonpointer==2.4
jupyter_client==8.6.0
jupyter_core==5.7.1
langchain==0.1.4
langchain-community==0.0.16
langchain-core==0.1.17
langchain-openai==0.0.5
langsmith==0.0.84
lxml==5.1.0
MarkupSafe==2.1.4
marshmallow==3.20.2
matplotlib-inline==0.1.6
mpmath==1.3.0
multidict==6.0.4
mypy-extensions==1.0.0
nest-asyncio==1.6.0
networkx==3.2.1
nltk==3.8.1
numpy==1.26.3
openai==1.10.0
packaging==23.2
parso==0.8.3
pexpect==4.9.0
phonenumbers==8.13.29
platformdirs==4.1.0
postgrest==0.15.0
prompt-toolkit==3.0.43
protobuf==4.25.2
psutil==5.9.8
psycopg2==2.9.9
ptyprocess==0.7.0
pure-eval==0.2.2
pydantic==2.6.0
pydantic_core==2.16.1
Pygments==2.17.2
PyJWT==2.8.0
PyPDF2==3.0.1
python-dateutil==2.8.2
python-docx==1.1.0
python-dotenv==1.0.1
python-stdnum==1.19
pytz==2023.4
PyYAML==6.0.1
pyzmq==25.1.2
realtime==1.0.2
regex==2023.12.25
requests==2.31.0
scikit-learn==1.4.0
scipy==1.12.0
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.25
stack-data==0.6.3
stanza==1.7.0
storage3==0.7.0
StrEnum==0.4.15
supabase==2.3.4
supafunc==0.3.3
sympy==1.12
tenacity==8.2.3
textblob==0.15.3
threadpoolctl==3.2.0
tiktoken==0.5.2
toml==0.10.2
torch==2.2.0
tornado==6.4
tqdm==4.66.1
traitlets==5.14.1
twilio==8.13.0
typing-inspect==0.9.0
typing_extensions==4.9.0
tzlocal==5.2
urllib3==2.1.0
wcwidth==0.2.13
websockets==11.0.3
Werkzeug==3.0.1
yarl==1.9.4
assemblyai